+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_rot13_env.php
@@ -44,7 +44,7 @@ class Symfony_DI_PhpDumper_Test_Rot13Parameters extends Container
     public function getRemovedIds(): array
     {
         return [
-            '.service_locator.ZZqL6HL' => true,
+            '.service_locator.PnIy5ic' => true,
             'Psr\\Container\\ContainerInterface' => true,
             'Symfony\\Component\\DependencyInjection\\ContainerInterface' => true,
         ];